pyproject.toml
[project]
name = "epitope-pipeline"
version = "0.1.0"
description = "B-cell epitope prediction pipeline using BepiPred, DiscoTope, and GraphBepi"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "torch>=2.11.0",
    "fair-esm>=2.0.0",
    "biopython>=1.86",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "plotly>=6.6.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "beautifulsoup4>=4.14.3",
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "ipykernel>=7.2.0",
    "papermill>=2.6",
]

[tool.uv.workspace]
members = [
    "B_cell_epitope_prediction",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/epitope_pipeline"]
